Move PageChooserPanel to wagtail.admin.panels.page_chooser_panel
